Refactor: Isolate Model and SvgModel
Isolate BaseModel.js from Model.js
Merge SVG methods into SvgModel.js
Rename Model.js to ThreeModel.js
Add onTransform on SvgModel to save transformation
Save on SVG Actions finished
